requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
adal=1.2.7=pypi_0
aiohttp=3.7.4.post0=pypi_0
alembic=1.4.1=pypi_0
applicationinsights=0.11.10=pypi_0
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
azure-common=1.1.27=pypi_0
azure-graphrbac=0.61.1=pypi_0
azure-mgmt-authorization=0.61.0=pypi_0
azure-mgmt-containerregistry=2.8.0=pypi_0
azure-mgmt-keyvault=2.2.0=pypi_0
azure-mgmt-resource=13.0.0=pypi_0
azure-mgmt-storage=11.2.0=pypi_0
azureml-core=1.28.0=pypi_0
azureml-mlflow=1.28.0.post1=pypi_0
azureml-telemetry=1.28.0=pypi_0
azureml-train=1.28.0=pypi_0
azureml-train-core=1.28.0=pypi_0
azureml-train-restclients-hyperdrive=1.28.0=pypi_0
backcall=0.2.0=pypi_0
backports-tempfile=1.0=pypi_0
backports-weakref=1.0.post1=pypi_0
blas=1.0=mkl
bleach=3.3.0=pypi_0
bokeh=2.3.2=py38h06a4308_0
boltons=21.0.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
configparser=5.0.2=pypi_0
contextlib2=0.6.0.post1=pypi_0
cryptography=3.4.7=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask=2021.5.1=pyhd3eb1b0_0
dask-core=2021.5.1=pyhd3eb1b0_0
databricks-cli=0.14.3=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pypi_0
defusedxml=0.7.1=pypi_0
distributed=2021.5.1=py38h06a4308_0
docker=4.4.4=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=pypi_0
expat=2.3.0=h2531618_2
ffmpeg=4.3=hf484d3e_0
flask=2.0.1=pypi_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.5.0=pypi_0
future=0.18.2=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.17=pypi_0
glib=2.68.1=h36276a3_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.30.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
gpytorch=1.5.0=pypi_0
greenlet=1.1.0=pypi_0
grpcio=1.37.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gunicorn=20.1.0=pypi_0
heapdict=1.0.1=py_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
importlib-metadata=3.10.0=py38h06a4308_0
intel-openmp=2021.2.0=h06a4308_610
ipdb=0.13.7=pypi_0
ipykernel=5.5.5=pypi_0
ipython=7.23.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
isodate=0.6.0=pypi_0
itsdangerous=2.0.1=pypi_0
jedi=0.18.0=pypi_0
jeepney=0.6.0=pypi_0
jinja2=3.0.1=pyhd3eb1b0_0
jmespath=0.10.0=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jsonpickle=2.0.0=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.7.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
locket=0.2.1=py38h06a4308_1
lz4-c=1.9.3=h2531618_0
mako=1.1.4=pypi_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.3.4=py38h06a4308_0
matplotlib-base=3.3.4=py38h62a2d02_0
matplotlib-inline=0.1.2=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
mlflow=1.17.0=pypi_0
msgpack-python=1.0.2=py38hff7bd54_1
msrest=0.6.21=pypi_0
msrestazure=0.6.4=pypi_0
multidict=5.1.0=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
ndg-httpsclient=0.5.1=pypi_0
nest-asyncio=1.5.1=pypi_0
nettle=3.7.2=hbbd107a_1
ninja=1.10.2=hff7bd54_1
notebook=6.4.0=pypi_0
numpy=1.20.2=py38h2d18471_0
numpy-base=1.20.2=py38hfae3a4d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py38h2531618_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
partd=1.2.0=pyhd3eb1b0_0
pathspec=0.8.1=pypi_0
pathtools=0.1.2=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.2.0=py38he98fc37_0
pip=21.1.3=py38h06a4308_0
plotly=4.14.3=pyhd3eb1b0_0
prometheus-client=0.10.1=pypi_0
prometheus-flask-exporter=0.18.2=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.18=pypi_0
protobuf=3.17.0=pypi_0
psutil=5.8.0=py38h27cfd23_1
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pydeprecate=0.3.0=pypi_0
pygments=2.9.0=pypi_0
pyjwt=2.1.0=pypi_0
pynvim=0.4.3=pypi_0
pyopenssl=20.0.1=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=pypi_0
python=3.8.8=hdb3f193_5
python-dateutil=2.8.1=pyhd3eb1b0_0
python-editor=1.0.4=pypi_0
pytorch=1.8.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.3.1=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py38h27cfd23_1
pyzmq=22.1.0=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=5.1.0=pypi_0
qtpy=1.9.0=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=py_2
rsa=4.7.2=pypi_0
ruamel-yaml=0.17.4=pypi_0
ruamel-yaml-clib=0.2.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.2=py38had2a1c9_1
secretstorage=3.3.1=pypi_0
send2trash=1.5.0=pypi_0
sentry-sdk=1.1.0=pypi_0
setuptools=52.0.0=py38h06a4308_0
shortuuid=1.0.1=pypi_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
smmap=4.0.0=pypi_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sqlalchemy=1.4.15=pypi_0
sqlite=3.35.4=hdfb4753_0
sqlparse=0.4.1=pypi_0
subprocess32=3.5.4=pypi_0
tabulate=0.8.9=pypi_0
tblib=1.7.0=py_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.10.0=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.8.1=py38
torchdiffeq=0.2.1=pypi_0
torchmetrics=0.3.2=pypi_0
torchsde=0.2.5=pypi_0
torchvision=0.9.1=py38_cu102
tornado=6.1=py38h27cfd23_0
tqdm=4.60.0=pypi_0
traitlets=5.0.5=pypi_0
trampoline=0.1.2=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pypi_0
wandb=0.10.31=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.0.0=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=pypi_0
zict=2.0.0=pyhd3eb1b0_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
